+++ src/Displayers/HtmlDisplayer.php
@@ -101,11 +101,12 @@ class HtmlDisplayer implements DisplayerInterface
     /**
      * Can we display the exception?
      *
-     * @param \Exception $exception
+     * @param \Exception $original
+     * @param \Exception $transformed
      *
      * @return bool
      */
-    public function canDisplay(Exception $exception)
+    public function canDisplay(Exception $original, Exception $transformed)
     {
         return true;
     }