+++ src/Filters/ContentTypeFilter.php
@@ -44,11 +44,12 @@ class ContentTypeFilter
      * Filter and return the displayers.
      *
      * @param \GrahamCampbell\Exceptions\Displayers\DisplayerInterface[] $displayers
-     * @param \Exception                                                 $exception
+     * @param \Exception                                                 $original
+     * @param \Exception                                                 $transformed
      *
      * @return \GrahamCampbell\Exceptions\Displayers\DisplayerInterface[]
      */
-    public function filter(array $displayers, Exception $exception)
+    public function filter(array $displayers, Exception $original, Exception $transformed)
     {
         foreach ($displayers as $index => $displayer) {
             if (!$this->request->accepts($displayer->contentType())) {