+++ src/Filters/FilterInterface.php
@@ -24,9 +24,10 @@ interface FilterInterface
      * Filter and return the displayers.
      *
      * @param \GrahamCampbell\Exceptions\Displayers\DisplayerInterface[] $displayers
-     * @param \Exception                                                 $exception
+     * @param \Exception                                                 $original
+     * @param \Exception                                                 $transformed
      *
      * @return \GrahamCampbell\Exceptions\Displayers\DisplayerInterface[]
      */
-    public function filter(array $displayers, Exception $exception);
+    public function filter(array $displayers, Exception $original, Exception $transformed);
 }