+++ src/Filters/VerboseFilter.php
@@ -44,11 +44,12 @@ class VerboseFilter
      * Filter and return the displayers.
      *
      * @param \GrahamCampbell\Exceptions\Displayers\DisplayerInterface[] $displayers
-     * @param \Exception                                                 $exception
+     * @param \Exception                                                 $original
+     * @param \Exception                                                 $transformed
      *
      * @return \GrahamCampbell\Exceptions\Displayers\DisplayerInterface[]
      */
-    public function filter(array $displayers, Exception $exception)
+    public function filter(array $displayers, Exception $original, Exception $transformed)
     {
         if ($this->config->get('app.debug', false) !== true) {
             foreach ($displayers as $index => $displayer) {