+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGE LOG
 
 * Associated uuids with exceptions
 * Added default displayer config
+* Resolve all the config earlier
 
 
 ## V3.2 (06/07/2015)